remover chave pix integrado com BCB

+++ src/main/kotlin/br/com/zup/academy/pix/remover/RemoveChavePixService.kt
@@ -1,8 +1,10 @@
 package br.com.zup.academy.pix.remover
 
+import br.com.zup.academy.integracao.banco_central.BancoCentralClient
 import br.com.zup.academy.pix.ChavePixNaoEncontradaException
 import br.com.zup.academy.pix.ChavePixRepository
 import br.com.zup.academy.pix.toUUID
+import io.micronaut.http.HttpStatus
 import io.micronaut.validation.Validated
 import javax.inject.Singleton
 import javax.validation.Valid
@@ -10,7 +12,8 @@ import javax.validation.Valid
 @Singleton
 @Validated
 class RemoveChavePixService(
-    val repository: ChavePixRepository
+    val repository: ChavePixRepository,
+    val bancoCentralClient: BancoCentralClient
 ) {
 
     fun remover(@Valid removeChavePixForm: RemoveChavePixForm){
@@ -23,6 +26,11 @@ class RemoveChavePixService(
 
         if(chavePix.idTitular != removeChavePixForm.idTitular?.toUUID()){
             throw IllegalArgumentException("A chave pode ser removida somente pelo seu dono")
+        }
+
+        val responseBcb = bancoCentralClient.deleta(key = chavePix.chave)
+        if ((responseBcb.status != HttpStatus.OK) && (responseBcb.status != HttpStatus.NOT_FOUND)) {
+            throw IllegalStateException("Erro ao remover chave Pix no Banco Central do Brasil (BCB)")
         }
 
         repository.delete(chavePix)
